export/discord-bot/requirements.txt
discord.py==2.5.2
wavelink==3.2.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
psutil==5.9.6
python-dateutil==2.8.2
Pillow>=9.0.0
PyNaCl>=1.5.0